requirements.txt
Flask==2.3.2
flask_sqlalchemy
sqlalchemy_utils
pymysql
python-dotenv
Authlib
Flask-OAuthlib
passlib[bcrypt]
